patchcore/pyproject.toml
[project]
name = "patchcore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "fpdf2>=2.8.3",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
]
